Add a Component that will Reload Tabs if needed (#43)
This add's a class that will listen to changes in proxyHandler. It will
reload active impacted tabs by the change. Not active tabs will be
discarded, and reloaded on the next activation. (As we probably dont
want to reload all of my 100 jira tabs that i have not been touching).
Also this will not reload "audible" tabs, so that a "background youtube
tab" is not reloaded if you change another youtube rule.