Implement Linux support via UNIX socket proxy (#93)
This is a companion PR to mozilla-mobile/mozilla-vpn-client#9952 that
implements Linux support by connecting to the SOCKS proxy as a systemd
service. Basically all we do here is check if the platform is `linux`
and set the bypass proxy to `{type: "socks", host:
"file:/var/run/mozillavpn.proxy", port: 1234}` The port number here
isn't used, but it seems to be required in order to be parsed as a valid
proxy.

Update! Now it supports connecting to the Windows proxy service too.
Similar logic, we just check the platform and generate a fixed proxy
config.