FXVPN-293: Improve Unsupported Platform message visibility (#129)
This PR: 
- Sets the `webExtension` flag to false by default. 
- This prevents folks on unsupported platforms with older versions of
the client (which predate the `webExtension` flag) from being shown the
the "Update now" panel instead of the "Unsupported Platform" panel.
Tested this with a v2.24.2 on mac.
- Identifies the OS with `await browser.runtime.getPlatformInfo()` and
uses this to conditionally show the unsupported platform panel.
- Preserves the ability for people on unsupported platforms to use the
extension by enabling the `webExtension` flag.
- Requests the `featurelist` when the popup is opened to catch changes
to the `webExtension` flag

A side-effect of this PR is that Windows users who flag off
`webExtension` flag will not see any changes in the extension. Given the
localproxy will now run on Windows without a flag I think this is
acceptable.