FXVPN-337 Return of the loading screen (#180)
If you look closely on the Screen Recording Shared with QA, you see the
Red dot on the Icon not appearing for quite a while.
This indicates firefox takes it's time inizializing the background
script.

You can easily replicate that by dropping ` await new Promise((r) =>
setTimeout(r, Math.round(Math.random() * 5000)));` somewhere in
`main::init()`.

Currently if we open the message port before the background script is
done inizializing we are screaming into the void.
Sooo, let's allow ipc.js to close and re-open ports if there is noone
answering on the other side.

For that ofc we need to show "Something" so i'm bringing back the
loading screen we added and then removed. 🙈

Here is a demo with a random delay. 


![output](https://github.com/user-attachments/assets/806c9150-c0be-4555-a5ab-27a48d92b9a9)