Dont use slots for the conditional view (#196)
Soo one thing, given we currently have all messages in light dom, which
means in a futs that they could be displayed.
Let's instead only create them on demand when we have the info to
collect the correct screen. Otherwise we should render nothing, as we
now have a placeholder.