Fix FXVPN-409 (#211)
This is a small tweak to the work of #205 . This checks that a browser
proxy settings is !-- "none" _and_ that a proxy URL has been provided.
Otherwise it is entirely possible for someone to choose 'Manual proxy
configuration', enter no URL, and then experience FXVPN-409 (which is I
suspect what happened in Raluca's case).

I will add tests in the AM :) 

Also: There is the matter of the 'Use system proxy settings' option in
browser proxy settings that I'm not sure how to handle or what to do
with. This PR would (I think) overwrite the system proxy setting (unless
a proxy URL is automagically attached to the proxy setting object) so
maybe that is a tweak that needs to be made.