FXVPN-440 Check Subscription Purchase Availability (#218)
Needs Final Content. 

The idea is simple, bedrock knows whether or not a client can buy the
vpn in the current locale.
Given bedrock also does not have an api endpoint for the
`vpn_available(request)` func, i'm simply loading the page in the
background and use their ui-test identifiers to check if the waitlist
will be shown.

That way we can save people quite some lifetime and hopefully some
frustration.


https://github.com/user-attachments/assets/9f5e9996-4015-4a9c-b55b-6c5e2cb67b07