Auto-Enable when opening a Private Browsing session (#239)
Add's code to enable the VPN when activating PBM. 

There are 2 Notable concepts to explain here to make this nicer to
review:

## Step 1: Refactor of ExtensionController

RN it assumed ownership of the controller state and had a tight
coupling.
i.e if the controller moved to `OnPartial` we'd activate too, because it
must have been us.
same thing, when deactivating in `OnPartial` we'd close the vpn. 
However other profiles exsist and with this pr even multiple instances
of this class.

So i'm moving the Extensionctroller State Machine in 3 Distinct Phases: 
-> Init: We're taking the first vpn-state (on/off) and mapping this to
(on/off) we will only do this once on startup
-> Unexpected-Changes: i.e if the client switched from off to OnPartial,
we should remain off. it wasnt us.
-> Expected-Changes: We just pressed toggle expecting something to
happen in the client, we'll follow state changes.

This should also fix a lot of the problem's we've seen with cross
profile things, eventhough i did not test for that yet.


## Step 2: Add a 2nd one. 

Now that ExtensionController is no longer conflicting with itself, we
can add a 2nd.
All consumers can then check if whatever they are doing is in context of
PBM or not and take the correct controller as input.


## Bonus: 
You can now cast any property into an async generator. 🥳
It's very neat as it auto-cancels the subscription once you cancel
iterating, no more manual cleanup (which we're not doing a good job of).
Also this allows us to use properties directly with
[asyncReplace](https://lit.dev/docs/api/directives/#asyncReplace) and
[asyncAppend](https://lit.dev/docs/api/directives/#asyncAppend) directly
in lit, without needing to subscribe and push values to local properties
:)


 





#### PR Dependency Tree


* **PR #239** 👈
  * **PR #241**
    * **PR #242**

This tree was auto-generated by
[Charcoal](https://github.com/danerwilliams/charcoal)