Perf: Check Deep Equality before passing data to UI (#249)
`Property` does not do an equality check, for good reasons, given we
know any value is frozen so it is another instance.
Therefore any call to set will always call all subscriptions.

The UI has everything subscribed, so any change to a prop will be passed
over IPC (meaning serializing & deserializing)
Now when we open the browser action we will request most data to be
refreshed
```
 connectedCallback() {
    super.connectedCallback();
    requestIdleCallback(() => {
      vpnController.postToApp("status");
    });
    requestIdleCallback(() => {
      vpnController.postToApp("servers");
    });
    requestIdleCallback(() => {
      vpnController.postToApp("featurelist");
    });
    requestIdleCallback(() => {
      vpnController.postToApp("interventions");
    });
    requestIdleCallback(() => {
      telemetry.record("main_screen");
    });
  }
  ```
  Which is fine but also means we're re-calculating all dependent data i.e when updating the controller state, we will update and re-calculate all properties that subscribed to that. which then again will push their changes to the UI. 
That is a LOT of work, and we might not need it. So let's de-dupe that on the vpncontroller.