FXVPN-363: adjust some CSS (#259)
~Video of the fix:

https://github.com/user-attachments/assets/a847a72a-b420-45a5-88b2-bbfdead94072~

Updated video of updated fix:
https://github.com/user-attachments/assets/1fe7cd60-b290-4424-a225-92f345f3722c

I believe both of these were conflicts caused by the popup's reduced
width when the scroll bar is visible. On an in-house page that doesn't
display the bottom server location stuff in the popup (like "thanks for
installing the extension" in that video above), the scroll bar isn't
needed and the problem doesn't currently exist on `main`. Same is true
when the butter bar isn't present - the height doesn't end up with a
scroll bar. The problem only exists on `main` when the popup includes
locations AND butter bar and thus needs the scroll bar.

(The divider in the purple box was also messed up - it was too long when
the browser had a scroll bar. This fix also addresses that piece.)

NOTE: The introduction of the scrollbar does cause the boxes to "jump"
to the left a bit. Let me know if we want a different treatment -
perhaps slowly transitioning or something similar.